+++ app/web/webpack.config.js
@@ -15,7 +15,7 @@ module.exports = {
     ],
   },
   optimization: {
-    minimize: true,
+    minimize: false,
   },
   devtool: "cheap-module-source-map",
   plugins: [
