+++ package.json
@@ -28,7 +28,6 @@
     "ts-loader": "^9.4.2",
     "ts-node": "^10.9.1",
     "tsc": "^2.0.4",
-    "typescript": "^4.9.4",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1",
     "webpack-dev-server": "^4.11.1",
