+++ pkg/gamertc/src/components/Gold.ts
