+++ pkg/gamertc/src/components/Tier.ts
