+++ pkg/gamertc/src/components/Traits.ts
