+++ pkg/gamertc/src/core/ecs/Component.ts
