+++ pkg/gamertc/src/core/ecs/System.ts
