+++ pkg/gamertc/src/core/net/Compression.ts
