+++ pkg/gamertc/src/core/util/Util.ts
