+++ pkg/gamertc/src/entities/Board.ts
