+++ pkg/gamertc/src/entities/Player.ts
