+++ pkg/gamertc/src/entities/Skelly.ts
