+++ pkg/gamertc/src/entities/Tile.ts
