+++ pkg/gamertc/src/entities/renderables/Button.ts
@@ -1,5 +1,5 @@
 import { Graphics, Text } from "pixi.js";
-import { Renderable, RenderableProps } from "./Renderable";
+import { Renderable, RenderableProps } from "../../core/ecs/Renderable";
 
 export type ButtonProps = RenderableProps & {
   dims: {w: number, textX: number, textY: number},
