+++ pkg/gamertc/src/entities/renderables/readme.md
