+++ pkg/gamertc/src/main.ts
@@ -1,22 +1,22 @@
-export * from "./net/NetManager";
-export * from "./game/Game";
-export * from "./ecs/Entity";
-export * from "./ecs/Component";
-export * from "./ecs/System";
-export * from "./graphics/Renderer";
-export * from "./graphics/Renderable";
-export * from "./graphics/Button";
-export * from "./graphics/Character";
-export * from "./graphics/Floor";
-export * from "./graphics/TextBox";
-export * from "./util/Util";
-export * from "./ecs/components/Health";
-export * from "./ecs/entities/Spaceship";
-export * from "./ecs/entities/Skelly";
-export * from "./ecs/entities/Unit";
-export * from "./ecs/entities/Player";
-export * from "./ecs/components/Experience";
-export * from "./ecs/components/Gold";
-export * from "./ecs/components/Name";
-export * from "./ecs/components/Tier";
-export * from "./ecs/components/Traits";
+export * from "./core/util/Util";
+export * from "./core/net/NetManager";
+export * from "./core/ecs/System";
+export * from "./core/ecs/Entity";
+export * from "./core/ecs/Component";
+export * from "./core/game/Game";
+export * from "./core/game/Renderer";
+export * from "./core/ecs/Renderable";
+export * from "./components/Traits";
+export * from "./components/Tier";
+export * from "./components/Name";
+export * from "./components/Health";
+export * from "./components/Gold";
+export * from "./components/Experience";
+export * from "./entities/renderables/TextBox";
+export * from "./entities/renderables/Floor";
+export * from "./entities/renderables/Character";
+export * from "./entities/renderables/Button";
+export * from "./entities/Unit";
+export * from "./entities/Spaceship";
+export * from "./entities/Skelly";
+export * from "./entities/Player";
