+++ pkg/gamertc/src/systems/DamageSystem.ts
