+++ pkg/games/pong/package.json
@@ -7,11 +7,11 @@
     "build": "yarn tsc",
     "start": "yarn ts-node src/main.ts"
   },
+  "dependencies": {
+    "@piggo-legends/gamertc": "1.0.0"
+  },
   "devDependencies": {
     "ts-node": "10.9.1",
     "typescript": "^4.9.5"
-  },
-  "dependencies": {
-    "@piggo-legends/gamertc": "1.0.0"
   }
 }
